pyproject.toml
[project]
name = "EnvTry"
version = "0.1.0"
description = ""
authors = [
    {name = "KabaevNV",email = "nkkab2@mail.ru"}
]
readme = "README.md"
requires-python = "3.13"
dependencies = [
    "fastapi (>=0.116.2,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "databases[asyncpg] (>=0.9.0,<0.10.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "bcrypt (==4.0.1)", # Версия для исключения ошибки с passlib
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
